refactor: move compile_kql_query to Engines/modules/systems/kql.py
KQL query compilation is a query-language concern, not a deployment
concern. Move it out of deployment.py into a dedicated kql module
under Engines/modules/systems/ alongside the other system-specific
modules.
